Add files via upload
A Python 3.0 Class to solve heat equation in 1 dimension. The folder has a main called main2 - which is the code to execute - and a class called myFEM2 - which is the header. By default the code will solve the 6.8 exercise of the second edition of Partial Differential Equation by Mark Gochenback. You can personalize what you want from the spatial mesh and the coefficients of the equation (this can be done by terminal) to the external source of heat (this must be done acting on the code). Please execute firstly as it is just to know how it works.
< Tess >